+++ apps/web/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @bombfarm/web
 
+## 0.19.1
+
+### Patch Changes
+
+- Updated dependencies [1ef139c]
+- Updated dependencies [1ef139c]
+  - @bombfarm/pricing@0.3.0
+  - @bombfarm/ui@0.13.1
+  - @bombfarm/domain@1.1.1
+  - @bombfarm/account@0.2.5
+  - @bombfarm/farm@1.1.1
+  - @bombfarm/game-art@0.5.1
+  - @bombfarm/hero@0.2.1
+  - @bombfarm/team-plan@0.1.1
+
 ## 0.19.0
 
 ### Minor Changes